In-progress work on making a "simple" cross-platform multiple-variant file-auto-generating auto-library-detecting singing-and-dancing scons build to rule them all.
Notes:
- yacc makes me want to yack. :(  

This is not working right now because of the complex way the parser was being created in the make build where it generated it to a non-standard 
filename and then moved it to another non-standard filename and then added it on as a file.

- Now need to start getting all the defined symbols in from options (like gv-enable) and others.  First I need to get a list of what all these 
flags are though so I can generate them and put then into OSGConfigured or something similar.

- This build uses scons-addons which is a project I created a long time ago and have been using on many projects.  Once the build is stable we can 
always import a stable version of use svn:external to pull it from it's svn repo.



git-svn-id: https://opensg.vrsource.org/svn/trunk@5 4683daeb-ad0f-0410-a623-93161e962ae5