WORKAROUND: This works around an inconsistent seg fault we were seeing when stressing the system heavily with load and deletes of full models.
The problem that we tracked down is that somehow a ContainerChangedEntry seemed to be processed after being cleared.  (ie. it was processed with container id of 
0 and NULL pointers).  We found that by removing the pooling of allocated container change entries the code would run without error.  This commit simplifies the 
pooling by allocating new ContainerChangeEntries every time they are needed and simply storing them in a vector.  When the change list is cleared, we iterate 
through this vector and deallocate the entries.  It is possible that this implementation may perform worse then the previous implementation but it does run 
without error for us.

This commit will need to be review in detail and will most likely either be improved to perform better or removed if the bug in the previous implementation is 
found.


git-svn-id: https://opensg.vrsource.org/svn/trunk@600 4683daeb-ad0f-0410-a623-93161e962ae5