Further enhancements to the memory debugging support.
This adds a "freelist" for field containers.  The idea is that when an FC is freed, there could still be FCPtr's referencing it 
(dangling) and we want to detect their use.  This is difficult though because libc may allocate new memory where that previous fc had 
been and then we don't have a way to store a flag to show that the fc has been deallocated (see previous commit).  This commit helps 
by delaying the freeing of memory for an fc for a while.  This way libc will not have the memory available to reallocate and instead 
we will be able to detect invalid access's for a longer time.  The number of fc's kept around is controlled by a limit (1 million for 
now).

Note:  unfortunately this commit includes some whitespace changes to (smart editor :( ), so if you view the diff, please remember to 
disable whitespace differences so you can see a cleaner view of what has changed. 


git-svn-id: https://opensg.vrsource.org/svn/trunk@638 4683daeb-ad0f-0410-a623-93161e962ae5