Add more memory checking code.
This change adds support for a table of metadata for fc's that have been allocated and freed.  Right now what this stores is the callstack 
at allocation time and deallocation time.  When an access to a destroyed fc is detected, the system will print the allocation stack, the 
deallocation stack, and the stack at the time of access.

This change also move the fcbaseptr subReference call to the .cpp file since it is highly unlikely it could be inlined.


git-svn-id: https://opensg.vrsource.org/svn/trunk@642 4683daeb-ad0f-0410-a623-93161e962ae5