Major refactoring of how the geo prop member functions are defined.
This change moves the definition of many of the template methods of the typedGeoXXX classes into the associated .cpp files.  This change 
was needed in part to force the compiler to generate strong symbols for these methods in a single shared object.

In the process of making this change I cleaned up the macros used for template class definition: remove unused macros, rename macros more 
accurately, expand the macros to support inline and non-inline use, and added macros to create full specializations.  There is more work 
to be done here, but this is a good start and trys to only change the macros that were directly used by the geo prop code.


git-svn-id: https://opensg.vrsource.org/svn/trunk@705 4683daeb-ad0f-0410-a623-93161e962ae5