[PATCH 02/11] Work around models that have texcoords in the wrong slot.
Add code that detects the case that we do not have a texture coordinates
in the first slot. In that case we search the remaining slots looking
for the first property that contains texture coordinates. If we find
texture coordinates, we move the texture coordinate properties into
the first slot so that they will work with fixed function OpenGL.

Signed-off-by: gerrit <vossg@vossg.org>