Utilize const pointers.
This is done in an effort to get the compiler to help identify places where
the code might modify source data. This does not technically fix anything.
Instead, it serves to give me a sanity check regarding the extraction of data
being copied from GIF image to an OSG::Image.