Fixed OSG::getenv() on Windows.
Do not define OSG_STATIC_MEMEBER_NEEDS_COPY_ASIGN_INIT when using Visual C++.
I suspect that this was necessary with Visual C++ 7.1, but it definitely isn't
needed with 9.0. It looks like type object registration has changed in a way
that causes registration to occur on OSG::TypeBase construction, and if that
happens twice (once for a temporary and once for the actual static data
member), then that leads to crashes when the type factory is initialized.