Add MultiProperty* and MultiLightChunk and examples
Patch by Johannes Brunen

MultiLightChunk
This one allows the specification of arbitrarily many lights and it
setups a shader storage buffer block that allows to access these lights
from the shader code. It allows the specification of the light types
(Point, Spot,...) and also of the details of the light's definition.
For instance it supports simple range based lights along to standard GL
lights. The first is specified by its color, intensity and range
parameter, while the latter are specified by the ambient, diffuse
and specular colored intensities and the attenuation factors.
However, also mixed specifications of these light types are possible.
Additionally, also the support for cinematic lights (ueber-lights) is
provided. It is also possible to request the addition of transform
matrices and/or their inverses for the spaces
involved (LS -> WS, LS -> VS) or the light space perspective matrix
and/or their inverse. Along to this chunk I have written three examples
to illustrate and test its usage and intend. These examples are named
'multirangelight.cpp', 'multistandardlight.cpp' and 'multicinemalight.cpp'.
Because of the simulation purpose they are a little elongated but the
actual code to be written for the lights is quite short.

MultiPropertySSBOChunk
MultiPropertyUBOChunk

These are convenient chunks that allow the setup of uniform buffers or
shader storage buffers, respectively, but with a simplified usage
pattern. The simpler interface comes with the price that they are
restricted with respect to what buffer structure can be represented
by them. However, they are for instance perfectly fine for the common
case to provide material data or simple state information into the
shader. Along with these I have written to test examples
'multipropertyssbochunk_test.cpp' and 'multipropertyubochunk_test.cpp'.
The previously mentioned examples 'multirangelight.cpp',
'multistandardlight.cpp' and 'multicinemalight.cpp' makes proper use of
both of them.